Avoid indexing bitvec via `[]` as this generates suboptimal assembly
`bit-vec` uses a `static TRUE/FALSE` trick to make indexing via
ops::Index possible. The trait requires to return a reference to
`Self::Output`. This is not really possible with a bit vector. The
interesting solution is to create two static bools (one true, one
false) and return a reference to those two instead. I noticed in
a bigger project using `stable-vec` that this leads to very bad
assembly (an additional branch and all in all more instructions).

This commit fixes that for the most important places. However,
this is only a quick and dirty solution. The crate will be rewritten
from scratch soon-ish. Then, there will be a ton more performance
tuning.