Move `from_vec` impl to `StableVec`
We can't just simply take the allocation of the `Vec` and treat it as
our own. So we have to allocate anyway. So there is not a huge benefit
from letting each core implement this. If we figure out this is in fact
a performance bottleneck, we can still optimize this function later.