Rename `StableVec` to `StableVecFacade` and add some type aliases
I think this is better than having a default type argument. That type
argument needed to be specified sometimes, so that's fairly annoying.
The only disadvantage of this is that users won't see all the methods
for `StableVec` but have to click on `StableVecFacade`. But I don't
think this is a problem.