Remove `from_vec`
This method cannot be implemented in a particularly efficient way
compared to simply using `from_iter`. Initially, I thought that this
method could work without allocating memory and basically doing
nothing. But that's rarely the case. To make the public API smaller,
this method is removed due to its questionable value.