Rename and change semantics of methods searching for a filled slot
`next_index_from` -> `first_filled_slot_from`
`prev_index_from` -> `first_filled_slot_below`

More importantly, the semantics of `first_filled_slot_below` changed in
that the `start` index is not included in the search. This makes sense
for a couple of reasons. Compare standard ranges: the upper limit is
excluded. So this makes sense for these methods as well. For example,
calling the method with 0 now will always yield `None` which is really
convenient for iterators and loops.